LPS-166968 Simplify, inline the method resolveFieldTypesModules as it is only used by one place
